Added Panel Identification using Timed Polling
Added function to continuously poll the active window, and set it as the selected window for editing the GUI on. As clicking on the GUI builder window automatically reset's itself as the active window, this is now ignored. And only panels added to gui_panels dictionary are recognized as selectable windows. Also added an automatic naming scheme for the panels.