Component Configuration Class for Editor
Added a component configurator class for the editor panel. It contains descriptions of each component, its parameters and the type of data in each parameter. This is then mapped to builder functions for each parameter configurator within the editor panel. The callbacks for these configurators are assigned to functions that set the specified attribute/attributes.