Testing Window Polling using a Global Listner
The previous method for determining the active window was to use an interrupt driven thread to poll the active window every second. However, this was wasteful and had a variable update latency (upto 1000ms) between the click and the update. However, we can now implement an global mouse click listener which will query the active window only when there is a mouse click event. This might seem more performance consuming, but overall it will be both more performant and will have near zero update latency.