Project GUI Data Format for Saving
This file is to test and finalize the data format for saving, restoring and building the final GUI. The format specification is still being finalized. We will be using pickled dictionaries and lists over JSON, XML or YAML. Because it is faster to serialize, doesn't have to be interpreted or converted to python objects after reading, and also doesn't need to be ported across applications, as it is specific to DPG and Python.