Added Window and Component Dictionaries
Added separate dictionaries for windows and components. Also, added methods for adding and removing components from the census. The census dictionary marks the number of components used so far by the builder. It can help with default numbering. It does not represent the current count, as some components might have been deleted. But it represents the total number of components of a type that have ever been instantiated.