Saving Projects and Window Positions
Added function for auto-saving all project data to file when exiting the program. Also added a function for collecting and saving the position and size of all windows in the app to the save file. Saved project files are of extension type .guide, but they are just python pickle files.