Component Add Function
A function to add a specified component. The component is copied from its template in the component dictionary with default attributes to the current component data. A unique tag is also generated for it. If the component selected is anything except a window, the current selected container is set as the parent.