Component Add Completed
Added window handling for selected component types. Also invokes the builder and data manager to create the specified component structurally with its attributes, and add it to the project data repo. Finally, the selection text display is updated to reflect the new component. Finally, non-container type components are also added to the "click" selection listener registry.