Selection Handler
Created a selection handler which gets the tag of the item that was selected. If the selection component is a window specifically and the previously selected component is the same (i.e. the component was double clicked). Then the selected container cycles upwards through all the components parents for every subsequent non-interrupted click on the same component.