Move Position of Component
Added a function to handle the movement of a component positioned within a parent. This allows us to change the rendered order of children within a container. The direction (up or down) is an argument which rearranges the child list and then updates the list in the parent container accordingly.