Component Attribute Functions
Added a dictionary of attribute datatypes with the pseudo-datatype as key and the corresponding DPG instantiation function as value. This allows the editor to create the controls dynamically based on the associated attribute. The $LABEL and $DEFAULT value are provided from the component dictionary attribute list values and the $ATTR user-data value is set to the component dictionary attribute keys (reference names).